+++ backends/dolphinbackend/include/dolphinbackend/ComplexData.h
@@ -108,7 +108,7 @@ public:
 
     T& access(std::size_t linearIndex);
     const T& access(std::size_t linearIndex) const;
-    T& getValue(int x, int y, int z) ;//TODO
+    T& getValue(size_t x, size_t y, size_t z) ;//TODO
 
     inline T* getData() const { return data; }
     void setData(T* data) {this->data = data;}
